feat: helper for armenian date version
