+++ frontend/src/components/articleList/ArticleItem.tsx
@@ -1,0 +1,28 @@
+import { IArticle } from "../../models/interfaces/article";
+import Img from "../../shared/img/Img";
+import styles from "./article.module.scss";
+import MoreButton from "../../shared/moreButton";
+import { dateArmFormater } from "../../helpers/dateArmFormater";
+
+export function ArticleItem(props: IArticle) {
+  const { mainImage, url, title, publishedAt } = props;
+  const date = dateArmFormater(publishedAt);
+
+  return (
+    <div key={url} className={styles.article_card}>
+      <div className={styles.article_photo}>
+        <Img width={316} height={260} src={mainImage} alt={url} />
+      </div>
+
+      <div className={styles.name_container}>
+        <h3 className={styles.article_name}>{title}</h3>
+      </div>
+      <div className={styles.published_data}>
+        <span>{date}</span>
+      </div>
+      <div>
+        <MoreButton link={`/articles/${url}`} />
+      </div>
+    </div>
+  );
+}
