+++ frontend/src/components/articleList/article.module.scss
@@ -30,6 +30,6 @@
 
 .published_data {
   font-weight: 400;
-  font-size: 10px;
+  font-size: 12px;
   color: $primary;
 }
