+++ frontend/src/helpers/dateArmFormater.ts
@@ -1,0 +1,11 @@
+export const dateArmFormater = (data: string) => {
+  const date = new Date(data);
+
+  const formatted = date.toLocaleDateString("hy-AM", {
+    year: "numeric",
+    month: "long",
+    day: "numeric",
+  });
+
+  return formatted;
+};
